requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.13.1
fonttools==4.57.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.7
protobuf==6.30.2
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.27.0
setproctitle==1.3.5
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchvision==0.22.0+cu126
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.10
